% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinemk.R
\name{baselinemk}
\alias{baselinemk}
\title{Function to add baseline variable}
\usage{
baselinemk(dftotf, pat_id = "patient_id", time_var = "time")
}
\arguments{
\item{dftotf}{data frame to create baseline column.
The data frame must contain specified `pat_id` and
`time_var`.}

\item{pat_id}{string of characters representing the column
of patient id.}

\item{time_var}{string of characters representing the column
of time.}
}
\description{
`baseline` column is created based on the full dataframe
provided by the user. If the `time_var` column contains 
more than one potential baseline values (i.e. values <= 0)
then `baselinemk()` will arrange the times according to 
closest value to 0 then assign that as baseline = 1 
while the other potential values will be assigne -1.
}
