% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calplotDFgen.R
\name{calplotDFgen}
\alias{calplotDFgen}
\title{Generate DataFrame for Calibration plot (for Continuous Value Predictions)}
\usage{
calplotDFgen(observed, predicted, n = 10)
}
\arguments{
\item{observed}{Vector of observed values (Numeric type).}

\item{predicted}{Vector of predicted values (Numeric type).}

\item{n}{Number of splits for observed and predicted.
Default is 10.}
}
\value{
A lsit of data frames data frame components ..., or an object of class 
\code{ggplot} [??]
}
\description{
Generate DataFrame for Calibration plot (for Continuous Value Predictions)
}
