% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_person_data.R
\name{create_person_data}
\alias{create_person_data}
\title{Creates data on the person level}
\usage{
create_person_data(data, out_time = 90, outcome = "tug",
  time_var = "time", pat_id = "patient_id", list_bs = NULL)
}
\arguments{
\item{data}{A data frame in long form}

\item{out_time}{The target time for which prediction is needed. 
The number correspond to one of the knots.}

\item{outcome}{-}

\item{time_var}{-}

\item{pat_id}{-}

\item{list_bs}{A list of objects, each of class \code{brokenstick} 
or class \code{brokenstick_export}. The list components with the 
name specified in \code{outcome} is taken as the external 
broken stick model.}
}
\value{
A data frame with one row for each \code{pat_id}, with
a column named \code{yhat} containing the broken stick estimates
for each person at time point specified in \code{time_var}.
}
\description{
Creates data on the person level
}
\examples{
timedata <- create_testdata("smocc")
persondata <- create_person_data(data = timedata, out_time = 14/12,
    outcome = "hdc", time_var = "age", pat_id = "id", 
    list_bs = PMMSKNN::smocc_bs)
}
