% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_testdata.R
\name{create_testdata}
\alias{create_testdata}
\title{Create datasets for testing purposes}
\usage{
create_testdata(dataset_name)
}
\arguments{
\item{dataset_name}{A string. Currently, only 
\code{dataset_name = "smocc"} is supported.}
}
\value{
A data frame in the long form
}
\description{
Create datasets for testing purposes
}
\examples{
full <- create_testdata("smocc")
nrow(full)
}
