% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIdbyPerf.R
\name{extractIdbyPerf}
\alias{extractIdbyPerf}
\title{Extract ID by Bias: extracts ID value from loocv and percentile based on bias}
\usage{
extractIdbyPerf(test_proc, perc)
}
\arguments{
\item{test_proc}{object from \code{\link{preproc}}.}

\item{perc}{Floating point value indicating N/100 percentile ID to look for.}
}
\value{
A numeric value of patient id number
}
\description{
Extract ID by Bias: extracts ID value from loocv and percentile based on bias
}
