% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extvalid.R
\name{extvalid}
\alias{extvalid}
\title{Calculate external validation measures of bias, coverage and precision}
\usage{
extvalid(loocvobj, test_proc)
}
\arguments{
\item{loocvobj}{An object produced by \code{\link{loocv_function}}}

\item{test_proc}{An object produced by \code{\link{preproc}}}
}
\value{
A vector of length 3
}
\description{
Calculate external validation measures of bias, coverage and precision
}
