% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitrefgamlss.R
\name{fitrefgamlss}
\alias{fitrefgamlss}
\title{Fit reference GAMLSS on the full data}
\usage{
fitrefgamlss(dist_fam = NULL, train_post, test_post, outcome,
  time_elapsed, time_window = NULL, cs = FALSE, dfspec = NULL,
  d_f_m = 1.8, ptr_m = 1, d_f_s = 1.2, d_f_n = 1, d_f_t = 1, ...)
}
\arguments{
\item{dist_fam}{gamlss distribution specification}

\item{train_post}{- datasets, typically the \code{train_post} list component 
of the object produced by \code{\link{preproc}}.}

\item{test_post}{Idem, component \code{test_post}}

\item{outcome}{Name of the outcomes variable}

\item{time_elapsed}{- Name of the time variable. (type=string)}

\item{time_window}{- vector of numbers for `centiles.pred()`, `xvalues` argument}

\item{cs}{Logical that specifies whether to use cubic spline. 
The default \code{cs = FALSE} uses ...}

\item{dfspec}{Logical that specifies whether to the user sets
degrees of freedom (...not clear to me what it does, and how it 
interacts with the next set of arguments)}

\item{d_f_m}{... explain: arguments probably mean different things for different distributions. Might be preferable to package it with \code{dist_fam}}

\item{ptr_m}{-}

\item{d_f_s}{-}

\item{d_f_n}{-}

\item{d_f_t}{-}

\item{\dots}{Passed down to \code{gamlss}}
}
\value{
There are many possible return values
}
\description{
The function fits an initial GAMLSS distribution to the full data
in order to be used as a reference distribution for the following
patient like me GAMLSS fits
}
