% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listtodf.R
\name{listtodf}
\alias{listtodf}
\title{Convert Performance metrics into dataframe RMSE and COVERAGE}
\usage{
listtodf(listfile)
}
\arguments{
\item{listfile}{DOCUMENT}
}
\value{
DOCUMENT
}
\description{
Utility function to aggregate all nearest_n results 
for all people.
}
