% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_perf.R
\name{loocv_perf}
\alias{loocv_perf}
\title{Performance Metric Calculation Function: After LOOCV}
\usage{
loocv_perf(
  loocv_res,
  outcome,
  nearest_n = nearest_n,
  opt_cov = 0.5,
  perf_round_by = perf_round_by,
  train = TRUE
)
}
\arguments{
\item{loocv_res}{- The \code{loocv_res} components of an object 
produced by \code{\link{loocv_function}}}

\item{outcome}{- Name of the outcomes variable (type=string)}

\item{nearest_n}{- Numeric vector indicating the nearest number of matches 
to select from. The \code{loocv_function} will iterate through 
the number within the vector and select the number of matches that has
optimal bias, coverage, and precision.
The number of nearest_n should be within the range of number of individuals
That are in the data. For example if there are 500 individuals in the data,
one could do \code{nearest_n <- 10:100}}

\item{opt_cov}{- Float value indicating optimal coverage value used for `perfrank`.}

\item{perf_round_by}{- Integer value to indicate what decimal point will the performance values should be rounded by. Default is `perf_round_by = 4`, set to smaller value to be less coarse about ranking `nearest_n` values.}

\item{train}{- Logical indicating whether or not aggregating from training data or testing data. Deafult `TRUE`}
}
\value{
Returns a data frame containing the performance measures (\code{bias}, \code{rmse}, \code{zscore}, \code{coverage}, \code{precision}, dropped cases due to failure in predicting values (\code{dcase}), number of matches (\code{nearest_n}), and the normalized scores for each of the performances measures (\code{bsc},\code{rsc},\code{covsc},\code{presc},\code{zsc}, and finally the sum of the scores: \code{totscore})) based on the number of nearest neighbors mathced (specified as \code{nearest_n}).
}
\description{
Performance Metric Calculation Function: After LOOCV
}
