% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocvperf.R
\name{loocvperf}
\alias{loocvperf}
\title{Performance Metric Calculation Function: After LOOCV}
\usage{
loocvperf(loocv_res, train_o, bias = "raw", nearest_n = nearest_n)
}
\arguments{
\item{loocv_res}{The \code{loocv_res} components of an object 
produced by \code{\link{loocv_function}}}

\item{train_o}{-}

\item{bias}{- Column indicating which bias score to use. Default
is \code{'raw'}. Options: \code{'raw','rmse','zsc'}.}

\item{nearest_n}{- Sequence of the neaerest n values.}
}
\description{
Performance Metric Calculation Function: After LOOCV
}
