% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchIdExtract.R
\name{matchIdExtract}
\alias{matchIdExtract}
\title{Match ID Function: create vector of ID values}
\usage{
matchIdExtract(ord_data, mtype, n, m, i)
}
\arguments{
\item{ord_data}{dataset. Training data from \code{\link{preproc}}.}

\item{mtype}{Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}.}

\item{n}{Number of matches}

\item{m}{Number of repititions of obtaining \eqn{\dot{y}}}

\item{i}{Patient id indicator.}
}
\value{
A vector of patient id numbers
}
\description{
Match ID Function: create vector of ID values
}
