% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchIdExtractsknn.R
\name{matchIdExtractsknn}
\alias{matchIdExtractsknn}
\title{Match ID Function: create array of ID values closest matching to training or testing case}
\usage{
matchIdExtractsknn(data, train_test = "train__test", patid, formula)
}
\arguments{
\item{data}{Full dataset. Contains both Training and Testing data.}

\item{train_test}{Column name indicating whether individual belongs to
Training or Testing dataset. Train = 1, Test = 2 by default.}

\item{patid}{Column name indicating patient id.}

\item{formula}{Formula indicating the variables used for matching.
(e.g. \code{ ~ var1 + var2 + var3 }).}
}
\value{
An k x klist$test (or k x klist$train) array of patient id numbers
}
\description{
Match ID Function: create array of ID values closest matching to training or testing case
}
