% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchIdExtractTest.R
\name{matchIdExtractTest}
\alias{matchIdExtractTest}
\title{Match ID Function: create vector of (Training set) ID values for Test set cases}
\usage{
matchIdExtractTest(test_proc, mtype, i, n)
}
\arguments{
\item{test_proc}{object processed after using \code{\link{preproc}}.}

\item{mtype}{Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}.}

\item{i}{(Test) Patient id indicator.}

\item{n}{Number of matches}
}
\value{
A vector of patient id numbers
}
\description{
Match ID Function: create vector of (Training set) ID values for Test set cases
}
