% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchTestDataGen.R
\name{matchTestDataGen}
\alias{matchTestDataGen}
\title{Match Testing Data Generating Function: create matched trainig and testing pair data}
\usage{
matchTestDataGen(test_o = test_o, ord_data = ord_data, mtype = 1)
}
\arguments{
\item{test_o}{dataset. Testing data with patient_id ordered from \code{\link{preproc}}.}

\item{ord_data}{dataset. Training data with patient_id ordered from \code{\link{preproc}}.}

\item{mtype}{Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}.}
}
\value{
A vector of patient id numbers
}
\description{
Match Testing Data Generating Function: create matched trainig and testing pair data
}
