% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchTrainDataGen.R
\name{matchTrainDataGen}
\alias{matchTrainDataGen}
\title{Match Training Data Generating Function: create Data based on pmm}
\usage{
matchTrainDataGen(train_post = train_post, ord_data, mtype = 1, n,
  m = 5, i, time_elapsed = time_elapsed, outcome = outcome,
  seed = seed, matchprobweight)
}
\arguments{
\item{train_post}{dataset. Post operative training data from \code{\link{preproc}}.}

\item{ord_data}{dataset. Training data with patient_id ordered from \code{\link{preproc}}.}

\item{mtype}{Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}.}

\item{n}{Number of matches}

\item{m}{Number of repititions of obtaining \eqn{\dot{y}}}

\item{i}{Patient id indicator.}

\item{time_elapsed}{Name of the time variable. (type=string)}

\item{outcome}{Name of the outcomes variable}

\item{seed}{Seed for randomly selecting matches based on differences in predicted y}

\item{matchprobweight}{Logical that specifies whether to utilize probability sampling}
}
\value{
A vector of patient id numbers
}
\description{
Match Training Data Generating Function: create Data based on pmm
}
