% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{plm}
\alias{plm}
\title{Fit reference GAMLSS on the full data}
\usage{
plm(ref = ref, matchmodel = matchmodel)
}
\arguments{
\item{ref}{Reference model from the \code{\link{fitrefgamlss}}.}

\item{matchmodel}{- dataset generated from the \code{\link{matchTrainDataGen}}}
}
\value{
There are many possible return values
}
\description{
The function fits an initial GAMLSS distribution to the full data
in order to be used as a reference distribution for the following
patient like me GAMLSS fits
}
