% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmout.R
\name{plmout}
\alias{plmout}
\title{Obtaining output of updated patient level model predictions}
\usage{
plmout(plmr, perfout = perfout, i = i, time_window = time_window,
  mint, maxt, loocv = loocv, matchmodel = matchmodel,
  traintestmatchdf = traintestmatchdf, outcome = outcome,
  time_elapsed = time_elapsed, ord_data = ord_data,
  train_post = train_post, test_post = test_post,
  thresh_val = thresh_val)
}
\arguments{
\item{plmr}{Fitted gamlss object updated with patient level data
based on \code{\link{fitrefgamlss}} and \code{update()}.}

\item{perfout}{List of performance measures}

\item{i}{Patient id indicator.}

\item{time_window}{vector of numbers for `centiles.pred()`, `xvalues` argument}

\item{mint}{Numeric value indicating minimum time for predicting values}

\item{maxt}{Numeric value indicating maximum time for predicting values}

\item{loocv}{Whether or not to perform leave one out cross validation or just go straight to prediction. Should
have the userchoose value specified if `loocv=FALSE`}

\item{matchmodel}{Dataset generated from the \code{\link{matchTrainDataGen}}}

\item{traintestmatchdf}{Matched train test dataframe based on \code{\link{matchTestDataGen}}}

\item{outcome}{Name of the outcomes variable}

\item{time_elapsed}{Name of the time variable. (type=string)}

\item{ord_data}{Idem, component \code{train_o}}

\item{train_post}{Datasets, typically the \code{train_post} list component}

\item{test_post}{Idem, component \code{test_post}}

\item{thresh_val}{Numeric value indicating value of bias to ignore (not include in output)}
}
\value{
an array of performance measures for each individuals in the training and testing set
}
\description{
The function primarily takes the output of the updated reference gamlss()
and outputs results of the performance measures
}
