% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cal.R
\name{plot_cal}
\alias{plot_cal}
\title{Plot for statistical validity and calibration}
\usage{
plot_cal(plotobj, test_proc = test_proc, outcome = "tug",
  filt = FALSE, pred_sum = "mean", obs_dist = "median",
  loocv = TRUE, filter_exp = NULL, plot_cal_zscore = FALSE,
  wtotplot = FALSE, plotvals = FALSE, iqrfull = NULL, ...)
}
\arguments{
\item{plotobj}{An object produced by \code{\link{loocv_function}}}

\item{test_proc}{- Preprocessed object from \code{\link{preproc}}}

\item{outcome}{Name of the outcome variable}

\item{filt}{Logical indicating ...}

\item{pred_sum}{-}

\item{obs_dist}{-}

\item{loocv}{Logical indicating the type of plot: 
Model performance plot (if \code{loocv = TRUE}, default), 
or or calibration plot (if \code{loocv = FALSE}).}

\item{filter_exp}{-}

\item{plot_cal_zscore}{- Logical indicating whether to plot zscore calibration}

\item{wtotplot}{- Logical indicating wehter to include a weighted total score plot
that indicates the optimal n match based on equally weighting bias, coverage and 
precision}

\item{plotvals}{- Logical indicating whether to plot bias, coverage, and precision
values onto the calibration plot}

\item{iqrfull}{- Dataframe containing gamlss predictions which triggers the plotting of 
reference model prediction on the same plot as that of the patient like me 
predictions.
values onto the calibration plot}

\item{\dots}{Not used}
}
\value{
A list with components ..., or an object of class 
\code{ggplot} [??]
}
\description{
Creates two types of plots.
\enumerate{
  \item \emph{Model performance plots} showing average bias, 
  coverage, 50 percent PI width (mean IQR diefference), 
  and a combined score of these statistics, at various 
  choices for the number of matches.
  \item \emph{Calibration plots} showing the distribution of the 
  observed outcomes at several predicted values. Separate plots
  are made for the training and test data.}
}
