% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cal_lmer.R
\name{plot_cal_lmer}
\alias{plot_cal_lmer}
\title{Plot for calibration of LMM models}
\usage{
plot_cal_lmer(fit, train, test, outcome, perfObj = NULL, fitopt = 1,
  ...)
}
\arguments{
\item{fit}{An object produced by \code{lmer} function.}

\item{train}{A dataframe of training observations with \code{outcome} column
included.}

\item{test}{A dataframe of testing observations with \code{outcome} column
included.}

\item{outcome}{A column name (i.e. string type) indicating \code{outcome} column}

\item{perfObj}{An object from \code{\link{perfCalcLMM}}}

\item{fitopt}{Number indicating which LMM prediction method to use (mostly for CI).
1 = predictInterval() a; 2 = predictInterval() b; 3 = predictInterval() c; 
4 = bootMer() parametric; 5 = bootMer() semiparametric}

\item{\dots}{Used for specifying options in the predict() method}
}
\value{
A list with components ..., or an object of class 
\code{ggplot} [??]
}
\description{
Creates Calibration Plots:
\enumerate{
  \item \emph{Calibration plots} showing the distribution of the 
  observed outcomes at several predicted values. Separate plots
  are made for the training and test data.}
}
