% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_func.R
\name{plot_func}
\alias{plot_func}
\title{Plot function under plot_cal() function}
\usage{
plot_func(plotobj = plotobj, train = TRUE, test_proc = test_proc,
  filt = filt, iqrfull = iqrfull, pred_sum = pred_sum,
  obs_dist = obs_dist, outcome = outcome, filter_exp = NULL)
}
\arguments{
\item{plotobj}{An object produced by \code{\link{loocv_function}}}

\item{train}{-}

\item{test_proc}{-}

\item{filt}{Logical indicating ...}

\item{iqrfull}{- Dataframe containing gamlss predictions which triggers the plotting of 
reference model prediction on the same plot as that of the patient like me 
predictions.
values onto the calibration plot}

\item{pred_sum}{-}

\item{obs_dist}{-}

\item{outcome}{Name of the outcome variable}

\item{filter_exp}{- String. For filtering possible values of bias, precision, and coverage values that are out of range. (e.g. \code{"bias < 1.5"})}

\item{\dots}{Not used}
}
\value{
A list with components ..., or an object of class 
\code{ggplot} [??]
}
\description{
Creates two types of plots.
\enumerate{
  \item \emph{Model performance plots} showing average bias, 
  coverage, 50 percent PI width (mean IQR diefference), 
  and a combined score of these statistics, at various 
  choices for the number of matches.
  \item \emph{Calibration plots} showing the distribution of the 
  observed outcomes at several predicted values. Separate plots
  are made for the training and test data.}
}
