% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ind.R
\name{plot_ind}
\alias{plot_ind}
\title{Plot an individuals outcome trajectory using PMMSKNN-GAMLSS}
\usage{
plot_ind(
  test_proc = test_proc,
  outcome = "tug",
  idnum = 1,
  mtype = 1,
  n = 10,
  dist = BCCGo,
  df_m = 2,
  df_s = 1,
  df_n = 1,
  df_t = 1,
  xvalues = 3:200,
  xlab = "Time",
  ylab = "Outcome",
  ...
)
}
\arguments{
\item{test_proc}{Preprocessed object from \code{\link{preproc}}}

\item{outcome}{Name of the outcomes variable (type=string)}

\item{idnum}{Integer indicating the id of the individual}

\item{mtype}{Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}. \code{mtype} values are 
from \code{0} to \code{4}}

\item{n}{Integer representing number of matches.}

\item{dist}{Object of \code{\link{gamlss.dist}} specifying the distribution for gamlss}

\item{df_m}{Numeric value that specifies the degrees of freedom for the cubic spline specified for the mean parameter of the distribution specified according to \code{dist_fam}}

\item{df_s}{Numeric value that specifies the degrees of freedom for the cubic spline specified for the scale parameter of the distribution specified according to \code{dist_fam}}

\item{df_n}{Numeric value that specifies the degrees of freedom for the cubic spline specified for the shape parameter, specifically the \eqn{\nu} parameter, of the distribution specified according to \code{dist_fam}}

\item{df_t}{Numeric value that specifies the degrees of freedom for the cubic spline specified for the shape parameter, specifically the \eqn{\tau} parameter, of the distribution specified according to \code{dist_fam}}

\item{xvalues}{Vector of values for the prediction centile to be plotted (default: 3:200)}

\item{xlab}{String indicating x axis label for plot (default: "Time")}

\item{ylab}{String indicating y axis label for plot (default: "Outcome")}

\item{\dots}{Options to specify in the plotting of the percentile curves}
}
\value{
A list that contains 1) an object of class \code{ggplot} that outputs a gamlss predition curve for an individual in terms of their predicted values over time, 2) a data frame that contains the observations from the matched neighbors as well as the individual w idnum and 3) A data frame of centiles prediction results  (10th, 25th, 50th, 75th, and 90th)
}
\description{
Creates a single plot with the below specification:
\enumerate{
  \item \emph{Individuals data points} showing individual patient values 
  \item \emph{Matched indivduals' data points} showing the 
  data points from the nearest neighbors of the selected individual
  }
}
