% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NthP_plm.R
\name{plot_NthP_plm}
\alias{plot_NthP_plm}
\title{Plot Nth Percentile patient like me plots}
\usage{
plot_NthP_plm(plotobj, test_proc = test_proc, outcome = "tug",
  time_var = "time", nvec = c(0.2, 0.8), mtype = 1, n = 10,
  dist = BCCGo, df_m = 2, df_s = 1, df_n = 1, df_t = 1,
  xvalues = 3:200, ...)
}
\arguments{
\item{plotobj}{An object produced by \code{\link{loocv_function}}}

\item{test_proc}{- An object produced by \code{\link{preproc}}}

\item{outcome}{Name of the outcome variable}

\item{time_var}{Name of the time variable}

\item{nvec}{vector of two proportions (e.g. \code{nvec = c(0.2,0.8)}).
These values indicate the lower and upper n-th percentile/100 person to plot.}

\item{mtype}{- Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}.}

\item{n}{Integer representing number of matches.}

\item{dist}{object of gamlss.dist specifying the distribution for gamlss}

\item{df_m}{Number representing the degrees of freedome for the location parameter}

\item{df_s}{Number representing the degrees of freedome for the sigma parameter}

\item{df_n}{Number representing the degrees of freedome for the nu parameter}

\item{df_t}{Number representing the degrees of freedome for the tau parameter}

\item{xvalues}{Vector of values for the prediction centile to be plotted}

\item{\dots}{Not used}
}
\value{
A list with components ..., or an object of class 
\code{ggplot} [??]
}
\description{
Creates two plots.
\enumerate{
  \item \emph{Lower N_1-th plot} showing individual patient values 
  along with matched individuals and gamlss centile curves for the
  lower N-th vector value specified
  \item \emph{Upper N_2-th plot} showing individual patient values 
  along with matched individuals and gamlss centile curves for the
  upper N-th vector value specified
  }
}
