% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ref_plm.R
\name{plot_ref_plm}
\alias{plot_ref_plm}
\title{Plot reference and patient like me plot}
\usage{
plot_ref_plm(plotobj, test_proc = test_proc, outcome = "tug",
  time_var = "time", patnum = 219, dist = NO, df_m = 2, df_s = 1,
  df_n = 1, df_t = 1, xvalues = 3:200, ...)
}
\arguments{
\item{plotobj}{An object produced by \code{\link{loocv_function}}}

\item{test_proc}{- An object produced by \code{\link{preproc}}}

\item{outcome}{Name of the outcome variable}

\item{time_var}{Name of the time variable}

\item{patnum}{Integer value of the patient number to plot alongside
the population reference chart}

\item{dist}{object of gamlss.dist specifying the distribution for gamlss}

\item{df_m}{Number representing the degrees of freedome for the location parameter}

\item{df_s}{Number representing the degrees of freedome for the sigma parameter}

\item{df_n}{Number representing the degrees of freedome for the nu parameter}

\item{df_t}{Number representing the degrees of freedome for the tau parameter}

\item{xvalues}{Vector of values for the prediction centile to be plotted}

\item{\dots}{Not used}
}
\value{
A list with components ..., or an object of class 
\code{ggplot} [??]
}
\description{
Creates two plots.
\enumerate{
  \item \emph{Reference plot} showing reference population values 
  along with fitted gamlss centile curves 
  \item \emph{Patient like me plot} showing the distribution of the 
  outcome of matched patients along with the selected patients' predicted
  centile curve. 
  }
}
