% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMMSKNN-package.R
\docType{package}
\name{PMMSKNN-pkg}
\alias{PMMSKNN-pkg}
\alias{PMMSKNN-pkg-package}
\title{\pkg{PMMSKNN}: Curve Matching Validation Tools for Knee Surgery Outcomes.}
\description{
INSERT HERE DESCRIPTION OF THE PACKAGE
}
\note{
Development of this package was kindly supported by 
FUNDER under THIS GRANT NUMBER
}
\section{PMMSKNN functions}{

The main functions are:
\tabular{ll}{
  \code{preproc()}        \tab Preprocess the data\cr
  \code{loocv_function()} \tab Evaluate statistical properties\cr
  \code{plot_cal()}       \tab Create calibration plots}
}

\references{
Kittelson et al. (2019). \emph{Development and testing of a 
neighbors-based prediction for physical function after total 
knee arthroplasty}. In preparation.
}
\seealso{
\code{\link{brokenstick}}
}
