% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmmydotgen.R
\name{pmmydotgen}
\alias{pmmydotgen}
\title{PMM \eqn{\dot{Y}} Generating Function: Create Y dots through Predictive Mean Matching}
\usage{
pmmydotgen(df, formula, pat_id, m, seed = 1234, dftest = NULL)
}
\arguments{
\item{df}{training data}

\item{formula}{A string with the name of the splitting variables}

\item{pat_id}{Name of the variable with patient identification}

\item{m}{- Integer value indicating the number of replications to obtain
\eqn{\beta*} and \eqn{\dot{y}}.}

\item{seed}{Integer value indicating seed number for random generation
of \eqn{\sigma*} and \eqn{\beta*}}

\item{dftest}{test data.frame object}
}
\value{
A data frame that includes \eqn{\dot{y}}.
}
\description{
PMM \eqn{\dot{Y}} Generating Function: Create Y dots through Predictive Mean Matching
}
