% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predZsc.R
\name{predZsc}
\alias{predZsc}
\title{Predict Z-Score for Training and Testing Data}
\usage{
predZsc(plmr = plmr, traintestmatchdf = traintestmatchdf,
  matchmodel = matchmodel, ord_data = ord_data,
  train_post = train_post, test_post = test_post, loocv = loocv,
  time_elapsed = time_elapsed, outcome = outcome, i = i)
}
\arguments{
\item{plmr}{Fitted object from \code{\link{plm}}}

\item{traintestmatchdf}{Matched train test dataframe based on \code{\link{matchTestDataGen}}}

\item{matchmodel}{- dataset generated from the \code{\link{matchTrainDataGen}}}

\item{ord_data}{Idem, component \code{train_o}}

\item{train_post}{- datasets, typically the \code{train_post} list component 
of the object produced by \code{\link{preproc}}.}

\item{test_post}{Idem, component \code{test_post}}

\item{loocv}{- Logical indicating whether or not to predict for train or test set.}

\item{time_elapsed}{Name of the time variable. (type=string)}

\item{outcome}{Name of the outcomes variable}

\item{i}{Patient id indicator.
\code{TRUE} will result int zscores from the training set, otherwise zscores from test set
will be generated.}
}
\value{
There are many possible return values
}
\description{
The function that predicts Z-score values initial GAMLSS distribution to the full data
in order to be used as a reference distribution for the following
patient like me GAMLSS fits
}
