% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smocc.R
\docType{data}
\name{smocc_bs}
\alias{smocc_bs}
\title{SMOCC data, broken stick model export}
\format{A list with four components: 
\describe{
\item{hgt}{Object of class \code{brokenstick_export}}
\item{wgt}{Object of class \code{brokenstick_export}}
\item{hdc}{Object of class \code{brokenstick_export}}
\item{bmi}{Object of class \code{brokenstick_export}}
}}
\description{
This object is a list with three objects: \code{hgt}, \code{wgt}, \code{hdc} and \code{bmi}. Each list member was generated by \code{export()} applied to the fitted broken stick model. This information is useful for calculating broken stick estimates for newly entered growth data
}
\keyword{datasets}
