% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselinemk.R
\name{baselinemk}
\alias{baselinemk}
\title{Function to add baseline variable}
\usage{
baselinemk(df0, pat_id = "patient_id", time_var = "time")
}
\arguments{
\item{df0}{data frame to transform
(i.e. data frame where a new column - baseline column - will be added).
The data frame must contain specified `pat_id` and `time_var`.}

\item{pat_id}{string of characters representing the column
of patient id.}

\item{time_var}{string of characters representing the column
of time.}
}
\value{
A data frame that contains a new `baseline` column 
        indicating the baseline observation (i.e. row).
        `baseline` == 1: the baseline time point;
        `baseline` == -1: the time points before baseline;
        `baseline` == 0: the time points after baseline.
}
\description{
`baseline` column is created based on the full data frame
provided by the user. If the `time_var` column contains
more than one potential baseline values (i.e. values <= 0)
then `baselinemk()` will arrange the times according to
closest value to 0 then assign that as baseline == 1
while the other potential baseline values will be assigned == -1.
}
