% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calplotDFgen.R
\name{calplotDFgen}
\alias{calplotDFgen}
\title{Generate a DataFrame for producing a calibration plot (for Continuous Value Predictions)}
\usage{
calplotDFgen(observed, predicted, n = 10)
}
\arguments{
\item{observed}{Vector of observed values (Numeric type).}

\item{predicted}{Vector of predicted values (Numeric type).}

\item{n}{Number of splits for observed and predicted.
Default is 10.}
}
\value{
A data frame used to generate calibration plots (using \code{"ggplot"}) for training and testing data.
}
\description{
This is a utility function that takes two vectors of observed and predicted values
in order to generate a dataframe necessary for producing a calibration plot.
}
