% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIdbyBias.R
\name{extractIdbyBias}
\alias{extractIdbyBias}
\title{Extract ID by Bias: extracts ID value from loocv object based on the percentile of bias value of individuals}
\usage{
extractIdbyBias(loocvObj, perc)
}
\arguments{
\item{loocvObj}{object from \code{\link{loocv_function}}.}

\item{perc}{Floating point value indicating N/100 percentile ID to look for.}
}
\value{
A numeric value representing patient id number
}
\description{
`extractIdbyBias` is a utility function that returns one single id number based on the
`perc` (numeric) argument provided. This is useful to determine the n-th percentile person
in terms of the bias values obtained from the LOOCV process.
}
