% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIdbyPerf.R
\name{extractIdbyPerf}
\alias{extractIdbyPerf}
\title{Extract ID by Performance: extracts ID value from processed object from `preproc()` based on the percentile of bias value of individuals}
\usage{
extractIdbyPerf(test_proc, perc)
}
\arguments{
\item{test_proc}{object from \code{\link{preproc}}.}

\item{perc}{Floating point value indicating N/100 percentile ID to look for.}
}
\value{
A numeric value representing patient id number
}
\description{
`extractIdbyPerf` is a utility function that returns one single id number based on the 
`perc` (numeric) argument provided. This is useful to determine the n-th percentile person
in terms of the overall performance values 
(could be a combination of bias, coverage, and precision) obtained from the LOOCV process.
}
