% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocvperf.R
\name{loocvperf}
\alias{loocvperf}
\title{Performance Metric Calculation Function: After LOOCV}
\usage{
loocvperf(loocv_res, train_o, bias = "raw", nearest_n = nearest_n)
}
\arguments{
\item{loocv_res}{- The \code{loocv_res} components of an object
produced by \code{\link{loocv_function}}}

\item{train_o}{- Data frame. Specifically, training data with patient_id ordered based on fitted distal outcome value using predicted mean matching.
Generated using \code{\link{preproc}}. Example, \code{x <- preproc()},
then \code{x$train_o} would be used for this parameter.}

\item{bias}{- Column indicating which bias score to use. Default
is \code{'raw'}. Options: \code{'raw','rmse','zsc'}.}

\item{nearest_n}{- Numeric vector indicating the nearest number of matches
to select from. The \code{loocv_function} will iterate through
the number within the vector and select the number of matches that has
optimal bias, coverage, and precision.
The number of nearest_n should be within the range of number of individuals
That are in the data. For example if there are 500 individuals in the data,
one could do \code{nearest_n <- 10:100}}
}
\value{
Returns a data frame containing the performance measures (\code{bias}, \code{rmse}, \code{zscore}, \code{coverage}, \code{precision}, dropped cases due to failure in predicting values (\code{dcase}), number of matches (\code{nearest_n}), and the normalized scores for each of the performances measures (\code{bsc},\code{rsc},\code{covsc},\code{presc},\code{zsc}, and finally the sum of the scores: \code{totscore})) based on the number of nearest neighbors mathced (specified as \code{nearest_n}).
}
\description{
Performance Metric Calculation Function: After LOOCV
}
