% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchTestDataGen.R
\name{matchTestDataGen}
\alias{matchTestDataGen}
\title{Match Testing Data Generating Function: create matched trainig and testing pair data}
\usage{
matchTestDataGen(test_o = test_o, ord_data = ord_data, mtype = 1)
}
\arguments{
\item{test_o}{Data frame. Specifically, testing data with patient_id ordered based on fitted distal outcome value using predicted mean matching.
Generated using \code{\link{preproc}}. Example, \code{x <- preproc()}, 
then \code{x$test_o} would be used for this parameter.}

\item{ord_data}{Data frame. Specifically, training data with patient_id ordered based on fitted distal outcome value using predicted mean matching.
Generated using \code{\link{preproc}}. Example, \code{x <- preproc()}, 
then \code{x$train_o} would be used for this parameter.}

\item{mtype}{- Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}. \code{mtype} values are 
from \code{0} to \code{4}}
}
\value{
A vector of (patient) id numbers
}
\description{
Match Testing Data Generating Function: create matched trainig and testing pair data
}
