% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat_level_func.R
\name{pat_level_func}
\alias{pat_level_func}
\title{Calculate bias, coverage and CI width for neighbors-based prediction}
\usage{
pat_level_func(ref = ref, nearest, dist_fam = dist_fam,
  train_post = train_post, ord_data = ord_data,
  test_post = test_post, test_o = test_o,
  traintestmatchdf = traintestmatchdf, outcome = outcome,
  time_elapsed = time_elapsed, plot = plot,
  matchprobweight = matchprobweight, time_window = time_window,
  interval = interval, cs = cs, dfspec = dfspec, d_f_m = d_f_m,
  ptr_m = ptr_m, d_f_s = d_f_s, d_f_n = d_f_n, d_f_t = d_f_t,
  thresh_val = thresh_val, printtrace = printtrace,
  userchoose = userchoose, seed = seed, parallel = parallel,
  loocv = loocv, mtype = mtype, biasm = biasm, m = m)
}
\arguments{
\item{ref}{Fitted gamlss object from \code{\link{fitrefgamlss}}}

\item{nearest}{Numeric vector with number of matches per scenario}

\item{dist_fam}{gamlss distribution specification}

\item{train_post}{- datasets, typically the \code{train_post} list component 
of the object produced by \code{\link{preproc}}.}

\item{ord_data}{Idem, component \code{train_o}}

\item{test_post}{Idem, component \code{test_post}}

\item{test_o}{Idem, component \code{test_o}}

\item{traintestmatchdf}{Matched train test dataframe based on \code{\link{matchTestDataGen}}}

\item{outcome}{Name of the outcomes variable}

\item{time_elapsed}{- Name of the time variable. (type=string)}

\item{plot}{- Logical that specifies whether to output individual precision plots}

\item{matchprobweight}{- Logical that specifies whether to utilize probability sampling
when doing the mean matching. If TRUE, matches nearest n weighted on differnce in 
predicted outcome.}

\item{time_window}{- vector of numbers for `centiles.pred()`, `xvalues` argument}

\item{interval}{- Int value specifying the interval of individuals to skip}

\item{cs}{Logical that specifies whether to use cubic spline. 
The default \code{cs = FALSE} uses ...}

\item{dfspec}{Logical that specifies whether to the user sets
degrees of freedom (...not clear to me what it does, and how it 
interacts with the next set of arguments)}

\item{d_f_m}{... explain: arguments probably mean different things for different distributions. Might be preferable to package it with \code{dist_fam}}

\item{ptr_m}{-}

\item{d_f_s}{-}

\item{d_f_n}{-}

\item{d_f_t}{-}

\item{thresh_val}{-}

\item{printtrace}{- Logical that specifies printing of gamlss parameter estimation}

\item{userchoose}{- Int value indicating the choice that the user wants to use for the number of nearest matches}

\item{seed}{- Seed for probability sampling of the nearest matches}

\item{parallel}{- Number of cores used for parallel computing. Default = 1}

\item{loocv}{- Whether or not to perform leave one out cross validation or just go straight to prediction. Should
have the userchoose value specified if `loocv=FALSE`}

\item{mtype}{- Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}.}

\item{biasm}{- Column indicating which bias score to use for choosing optimal n. Default
is \code{'raw'}. Options: \code{'raw','rmse','zsc'}.}

\item{m}{Number of repititions of obtaining \eqn{\dot{y}}}
}
\value{
There are many possible return values
}
\description{
The function calculate three frequency-based parameters to 
demonstate statistical quality of the neighbors-based prediction.  #' The parameters are bias, coverage and the 50% prediction interval 
width.
}
