% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{plm}
\alias{plm}
\title{Fit reference GAMLSS on the full data}
\usage{
plm(ref = ref, matchmodel = matchmodel)
}
\arguments{
\item{ref}{- gamlss model object generated from using the \code{\link{fitrefgamlss}}. This would be a reference model fit to the whole population dataset.}

\item{matchmodel}{- dataset generated from the \code{\link{matchTrainDataGen}}}
}
\value{
A fitted gamlss object that is used within the \code{\link{loocv_function}} where the object is updated according to the closest \code{n} matches.
}
\description{
The function fits an initial GAMLSS distribution to the full data
in order to be used as a reference distribution for the following
patient like me GAMLSS fits
}
