% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmout.R
\name{plmout}
\alias{plmout}
\title{Obtaining output of updated patient level model predictions}
\usage{
plmout(
  plmr,
  perfout = perfout,
  i = i,
  time_window = time_window,
  mint,
  maxt,
  loocv = loocv,
  matchmodel = matchmodel,
  traintestmatchdf = traintestmatchdf,
  outcome = outcome,
  time_elapsed = time_elapsed,
  ord_data = ord_data,
  train_post = train_post,
  test_post = test_post,
  thresh_val = thresh_val
)
}
\arguments{
\item{plmr}{Fitted gamlss object updated with patient level data
based on \code{\link{fitrefgamlss}} and \code{update()}.}

\item{perfout}{List of performance measures to be filled.}

\item{i}{Patient id indicator (type=numeric).}

\item{time_window}{vector of numbers for `centiles.pred()`, `xvalues` argument}

\item{mint}{Numeric value indicating minimum time for predicting values}

\item{maxt}{Numeric value indicating maximum time for predicting values}

\item{loocv}{- Logical (\code{TRUE/FALSE}) that specifies whether 
or not to perform leave-one-out cross validation or just output 
predictions without hyperparameter tuning. If \code{loocv=FALSE}, then
users need to specify the value of the nearest_n}

\item{matchmodel}{Dataset generated from the \code{\link{matchTrainDataGen}}}

\item{traintestmatchdf}{Matched train test dataframe based on \code{\link{matchTestDataGen}}}

\item{outcome}{- Name of the outcomes variable (type=string)}

\item{time_elapsed}{Name of the time variable. (type=string)}

\item{ord_data}{Data frame. Specifically, training data with patient_id ordered based on fitted distal outcome value using predicted mean matching.
Generated using \code{\link{preproc}}. Example, \code{x <- preproc()}, 
then \code{x$train_o} would be used for this parameter.}

\item{train_post}{- Data frame that contains the post-baseline observations from the training dataset. Typically this would be the \code{train_post} list component that was generated from the \code{\link{preproc}} function}

\item{test_post}{Data frame that contains the post-baseline observations from the testing dataset. Typically this would be the \code{train_post} list component that was generated from the \code{\link{preproc}} function}

\item{thresh_val}{- Numeric value indicating value of bias to ignore (not include in output) in terms of the leave-one-out cross validation process. The default is set to \code{thresh_val = 10000}}
}
\value{
An array of performance measures for each individuals in the training and testing set
}
\description{
The function primarily takes the output of the updated reference gamlss()
and outputs results of the performance measures based on a fitted reference model (This happens within the \code{\link{pat_level_func}}).
}
