% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NthP_plm.R
\name{plot_NthP_plm}
\alias{plot_NthP_plm}
\title{Plot Nth Percentile patient like me plots using PMMSKNN-GAMLSS}
\usage{
plot_NthP_plm(
  test_proc = test_proc,
  outcome = "tug",
  nvec = c(0.2, 0.8),
  mtype = 1,
  n = 10,
  dist = BCCGo,
  df_m = 2,
  df_s = 1,
  df_n = 1,
  df_t = 1,
  xvalues = 3:200,
  chartname = "Patient Reference Chart",
  name = "Patient",
  ...
)
}
\arguments{
\item{test_proc}{- Preprocessed object from \code{\link{preproc}}}

\item{outcome}{- Name of the outcomes variable (type=string)}

\item{nvec}{- Vector of two proportions (e.g. \code{nvec = c(0.2,0.8)}).
These values indicate the lower and upper n-th percentile/100 person to plot.}

\item{mtype}{- Integer value indicating matching type. Default is set to 1 which follows the
matching of patients based on recommendation from \href{https://stefvanbuuren.name/fimd/sec-pmm.html}{van Buuren et al.}. \code{mtype} values are
from \code{0} to \code{4}}

\item{n}{- Integer representing number of matches.}

\item{dist}{- Object of \code{\link{gamlss.dist}} specifying the distribution for gamlss}

\item{df_m}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the mean parameter of the distribution specified according to \code{dist_fam}}

\item{df_s}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the scale parameter of the distribution specified according to \code{dist_fam}}

\item{df_n}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the shape parameter, specifically the \eqn{\nu} parameter, of the distribution specified according to \code{dist_fam}}

\item{df_t}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the shape parameter, specifically the \eqn{\tau} parameter, of the distribution specified according to \code{dist_fam}}

\item{xvalues}{- Vector of values for the prediction centile to be plotted}

\item{chartname}{- String indicating name of chart}

\item{name}{- String indicating name of object (default: "Patient")}

\item{\dots}{- Options to specify in the plotting of the percentile curves.}
}
\value{
An object of class \code{ggplot} that outputs a gamlss predition curve for 2 individuals in terms of their predicted values over time.
}
\description{
Creates two plots.
\enumerate{
  \item \emph{Lower N_1-th plot} showing individual patient values
  along with matched individuals and gamlss centile curves for the
  lower N-th vector value specified
  \item \emph{Upper N_2-th plot} showing individual patient values
  along with matched individuals and gamlss centile curves for the
  upper N-th vector value specified
  }
}
