% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ref_plm.R
\name{plot_ref_plm}
\alias{plot_ref_plm}
\title{Plot reference and personalized plot}
\usage{
plot_ref_plm(
  plotobj,
  test_proc = test_proc,
  outcome = "tug",
  time_var = "time",
  patnum = 219,
  dist = NO,
  df_m = 2,
  df_s = 1,
  df_n = 1,
  df_t = 1,
  xvalues = 3:200,
  ...
)
}
\arguments{
\item{plotobj}{- An object produced by \code{\link{loocv_function}}}

\item{test_proc}{- Preprocessed object from \code{\link{preproc}}}

\item{outcome}{- Name of the outcomes variable (type=string)}

\item{time_var}{- Name of the time variable. (type=string)}

\item{patnum}{- Integer value of the patient number to plot alongside
the population reference chart}

\item{dist}{- Object of \code{\link{gamlss.dist}} specifying the distribution for gamlss}

\item{df_m}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the mean parameter of the distribution specified according to \code{dist_fam}}

\item{df_s}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the scale parameter of the distribution specified according to \code{dist_fam}}

\item{df_n}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the shape parameter, specifically the \eqn{\nu} parameter, of the distribution specified according to \code{dist_fam}}

\item{df_t}{- Numeric value that specifies the degrees of freedom for the cubic spline specified for the shape parameter, specifically the \eqn{\tau} parameter, of the distribution specified according to \code{dist_fam}}

\item{xvalues}{- Vector of values for the prediction centile to be plotted}

\item{\dots}{- Options to specify in the \code{ggplot()} function used for plotting the reference and personalized plot}
}
\value{
An object of class \code{ggplot} that outputs a gamlss predition curve for the population and a personalized plot in terms of predicted values over time.
}
\description{
Creates two plots.
\enumerate{
  \item \emph{Reference plot} showing reference population values 
  along with fitted gamlss centile curves 
  \item \emph{Patient like me plot} showing the distribution of the 
  outcome of matched patients along with the selected patients' predicted
  centile curve. 
  }
}
