% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMMSKNN-package.R
\docType{package}
\name{PMMSKNN-pkg}
\alias{PMMSKNN-pkg}
\title{\pkg{PMMSKNN}: Sequential KNN Extended via Predicted Mean Matching.}
\description{
This package provides functions that allow personalized predictions of longitudinal outcome trajectory based on extending the sequential KNN algorithm via Predicted Mean Matching. Essentially individuals are matched according to a distal outcome of interest based on characteristics that explains the variation in the outcome. The trajectory of the matched population is used to generate predictions for each individuals.
}
\note{
Development of this package was kindly supported by
FUNDER under THIS GRANT NUMBER
}
\section{PMMSKNN functions}{

The main functions are:
\tabular{ll}{
  \code{preproc()}        \tab Preprocess the data\cr
  \code{loocv_function()} \tab Evaluate statistical properties\cr
  \code{plot_cal()}       \tab Create calibration plots}
}

\references{
Kittelson et al. (2019). \emph{Development and testing of a
neighbors-based prediction for physical function after total
knee arthroplasty}. In preparation.
}
\seealso{
\code{\link{brokenstick}}
}
