% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{PreProcess Function: create train/test filtering}
\usage{
preproc(dff, split_var = "train_test", trainval = 1, testval = 2,
  filter_exp = NULL, knots_exp = c(0, 14, 50, 90), out_time = 90,
  outcome = "tug", time_var = "time", pat_id = "patient_id",
  baseline_var = "baseline", varlist = NULL, pmmform = NULL,
  modelselect = FALSE, m = 5, ...)
}
\arguments{
\item{dff}{dataset}

\item{split_var}{A string with the name of the splitting variables}

\item{trainval}{-}

\item{testval}{-}

\item{filter_exp}{Expression can be of form \code{"time > 3"}. Default is \code{NULL}}

\item{knots_exp}{Numeric vector with break points for \code{brokenstick} model}

\item{out_time}{Single number value specifying the target time. Has to be one of 
\code{knots_exp}.}

\item{outcome}{Name of the outcomes variable}

\item{time_var}{Name of the time variable}

\item{pat_id}{Name of the variable with patient identification}

\item{baseline_var}{Name of the pre-op/post-op indicator variable: preop (baseline = 1); 
postop: (baseline = 0), or otherwise}

\item{varlist}{Names of additional variables for prediction. If not 
specified, all variables \code{dff} are predictors. 
Categorical variables may be factor or character.}

\item{pmmform}{-}

\item{modelselect}{A logical specifying ...}

\item{m}{Number of repititions of obtaining \eqn{\dot{y}}}

\item{\dots}{Not used.}
}
\value{
A list with six components.
}
\description{
PreProcess Function: create train/test filtering
}
